Re-architecture Phase 1: framework-free LegendSDLCException, complete legend-sdlc-shared
- Move LegendSDLCServerException down to legend-sdlc-shared as
  org.finos.legend.sdlc.error.LegendSDLCException, replacing the JAX-RS
  Response.Status field with a plain int status code (constructors default
  to 500 and validate* helpers to 400, exactly as before).
- Keep LegendSDLCServerException in legend-sdlc-server-shared as a
  @Deprecated subclass with the unchanged Status-typed API, so existing
  throw/catch/mapper sites are unaffected.
- Relocate StringTools and IOTools to org.finos.legend.sdlc.tools (per
  re-architecture.md par. 5, extracted classes leave server.* packages),
  leaving @Deprecated bridge subclasses at the old FQNs; update all
  in-repo references.
- Add TestLegendSDLCException; keep TestLegendSDLCServerException
  unchanged as the compatibility pin.
- Start docs/re-architecture-worklog.md recording Phase 1 decisions and
  the follow-ups Phase 2 must pick up.

Phase 1 of docs/re-architecture.md.
