+++ docs/re-architecture-worklog.md
@@ -1,0 +1,92 @@
+# Re-architecture Implementation Worklog
+
+Running log of the implementation of [`re-architecture.md`](re-architecture.md).
+One section per phase; within a phase, newest entries last. Records what moved,
+decisions taken where the plan allowed a choice, deviations from the plan, and
+bugs observed-but-preserved (Phases 2–3 are behavior-preserving; fixes are logged
+here and deferred). A fresh session should be able to resume from this file plus
+`git log` on the `reorg` branch.
+
+## Conventions adopted (apply to all phases)
+
+- **New package roots**: relocated classes leave `org.finos.legend.sdlc.server.*`
+  per re-architecture §5. Chosen so far: `org.finos.legend.sdlc.tools` (shared
+  utilities), `org.finos.legend.sdlc.error` (framework-free exceptions). These
+  follow the existing non-server precedents (`…sdlc.serialization`,
+  `…sdlc.domain.model.*`, `…sdlc.generation`).
+- **Bridge pattern for classes**: old FQN retained as a `@Deprecated` subclass of
+  the relocated class, in the module that historically shipped the old FQN. The
+  Javadoc says the bridge is retained *temporarily* — it does not promise a
+  specific release cycle (removal timing is coordinated per re-architecture §5).
+  Static utility methods are reachable through subclass names, so bridges for
+  utility classes are empty subclasses (relocated utility constructors were made
+  `protected` to permit this).
+- **Build/verify**: `mvn install javadoc:javadoc` on Java 25 (matches
+  `.github/workflows/build.yml`); checkstyle and `dependency:analyze`
+  (`failOnWarning=true`) are active, so a green build implies both.
+
+## Phase 1 — Foundations (completing `legend-sdlc-shared`)
+
+**Status: complete** (this commit).
+
+What moved:
+
+- `StringTools`, `IOTools` (+ tests): `org.finos.legend.sdlc.server.tools` →
+  `org.finos.legend.sdlc.tools`, staying in `legend-sdlc-shared`. `@Deprecated`
+  empty-subclass bridges left at the old FQNs in `legend-sdlc-shared` (that is
+  where the old FQNs currently ship on this branch). All in-repo references
+  (22 files across `legend-sdlc-project-files`, `-server`, `-server-fs`,
+  `-server-shared`) updated to the new package.
+- `LegendSDLCServerException` → **`org.finos.legend.sdlc.error.LegendSDLCException`**
+  in `legend-sdlc-shared`. The JAX-RS `Response.Status` field is replaced by a
+  plain `int statusCode` (`getStatusCode()`); defaults preserved exactly
+  (constructors default 500; `validate*` helpers default 400). Static `validate*`
+  helpers mirrored with `int` overloads.
+- `LegendSDLCServerException` remains in `legend-sdlc-server-shared` as a
+  `@Deprecated` subclass: identical constructor signatures (`Status`-typed,
+  null → 500), `getStatus()` now derived via `Status.fromStatusCode(getStatusCode())`
+  (exact round-trip — all constructors take the enum), and its own static
+  `validate*` helpers still constructing/throwing `LegendSDLCServerException` so
+  existing catch sites and the JAX-RS mapper are unaffected.
+  `TestLegendSDLCServerException` kept unchanged as the compat pin; new
+  `TestLegendSDLCException` covers the int-based API.
+
+Decisions / deviations:
+
+- **Constructor visibility loosened on relocated utilities**: `StringTools`'s
+  private constructor became `protected` (and `IOTools` gained an explicit
+  `protected` one, where it previously had an implicit public default) so the
+  deprecated bridges can extend them. Net API effect: `new IOTools()` is no
+  longer possible for outsiders — it was never meaningful.
+- **Serialization compatibility**: `LegendSDLCServerException` keeps its
+  `serialVersionUID`, but its serialized shape changed (status now an `int` in
+  the superclass instead of an enum field). Java-serialized instances of the old
+  class are not deserializable across this change. Judged acceptable: these are
+  server-side HTTP exceptions, never persisted; noted for the release notes.
+- **Split package `org.finos.legend.sdlc.server.tools` persists** for the
+  deprecation cycle: bridges in `legend-sdlc-shared`, plus 6 unrelated server
+  classes (`AuthenticationTools`, `BackgroundTaskProcessor`, …) in
+  `legend-sdlc-server`. Pre-existing condition; it ends when the bridges are
+  deleted. The server-only classes in that package are Phase 4/5 material.
+- **No exception mapper for `LegendSDLCException` yet**, deliberately: nothing
+  below the server throws it until Phase 2 moves code down. Phase 2 must (a) add
+  a JAX-RS mapper (or extend `BaseExceptionMapper`) for the new base type and
+  (b) audit `catch (LegendSDLCServerException …)` sites on any path that will
+  call relocated code, since relocated code will throw the base type, which those
+  catches do not match.
+
+Explicitly *not* done (deferred to their phases):
+
+- `legend-sdlc-server-shared`'s other contents (auth/session, `BaseServer`,
+  mappers, `BaseResource`) are server-ware and stay put (L6-adjacent).
+- The 6 `server.tools` classes in `legend-sdlc-server` (above).
+- Anything in `legend-sdlc-project-files` package renames (`server.project.*` →
+  `projectfiles`) — that is Phase 2 territory alongside the `SourceSpecification`
+  decision.
+
+## Phase 2 — Project structure extraction
+
+Not started. Inputs to carry in: seams S2 and R1 (re-architecture §6), the §5
+migration recipe + dual-keyed `ServiceLoader` lookup, the extension-SPI interface
+bridge, and the two Phase-1 notes above (mapper for `LegendSDLCException`; catch-site
+audit).
