+++ legend-sdlc-server-shared/src/main/java/org/finos/legend/sdlc/server/resources/BaseResource.java
@@ -16,7 +16,7 @@ package org.finos.legend.sdlc.server.resources;
 
 import org.finos.legend.sdlc.server.error.LegendSDLCServerException;
 import org.finos.legend.sdlc.server.monitoring.SDLCMetricsHandler;
-import org.finos.legend.sdlc.server.tools.StringTools;
+import org.finos.legend.sdlc.tools.StringTools;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
