+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/depot/api/DepotMetadataApi.java
@@ -30,7 +30,7 @@ import org.finos.legend.sdlc.server.depot.DepotServerInfo;
 import org.finos.legend.sdlc.server.depot.model.DepotProjectId;
 import org.finos.legend.sdlc.server.depot.model.DepotProjectVersion;
 import org.finos.legend.sdlc.server.error.LegendSDLCServerException;
-import org.finos.legend.sdlc.server.tools.StringTools;
+import org.finos.legend.sdlc.tools.StringTools;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
