+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/gitlab/api/GitLabApiWithFileAccess.java
@@ -51,8 +51,8 @@ import org.finos.legend.sdlc.server.project.ProjectFiles;
 import org.finos.legend.sdlc.server.project.ProjectPaths;
 import org.finos.legend.sdlc.server.project.ProjectStructure;
 import org.finos.legend.sdlc.server.tools.BackgroundTaskProcessor;
-import org.finos.legend.sdlc.server.tools.IOTools;
-import org.finos.legend.sdlc.server.tools.StringTools;
+import org.finos.legend.sdlc.tools.IOTools;
+import org.finos.legend.sdlc.tools.StringTools;
 import org.gitlab4j.api.CommitsApi;
 import org.gitlab4j.api.Constants;
 import org.gitlab4j.api.Constants.Encoding;
