+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/project/extension/DefaultProjectStructureExtension.java
@@ -24,7 +24,7 @@ import org.finos.legend.sdlc.server.project.ProjectFileAccessProvider;
 import org.finos.legend.sdlc.server.project.ProjectFileOperation;
 import org.finos.legend.sdlc.server.project.ProjectStructure;
 import org.finos.legend.sdlc.server.project.config.ProjectFileConfiguration;
-import org.finos.legend.sdlc.server.tools.StringTools;
+import org.finos.legend.sdlc.tools.StringTools;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
