+++ legend-sdlc-shared/src/test/java/org/finos/legend/sdlc/error/TestLegendSDLCException.java
@@ -1,0 +1,103 @@
+// Copyright 2026 Goldman Sachs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.finos.legend.sdlc.error;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.function.Function;
+import java.util.function.Predicate;
+
+public class TestLegendSDLCException
+{
+    @Test
+    public void testConstructorDefaults()
+    {
+        String message = "the quick brown fox jumped over the lazy dog";
+        Throwable cause = new RuntimeException("cause");
+
+        Assert.assertEquals(500, new LegendSDLCException(message).getStatusCode());
+        Assert.assertEquals(500, new LegendSDLCException(message, cause).getStatusCode());
+        Assert.assertEquals(409, new LegendSDLCException(message, 409).getStatusCode());
+        Assert.assertEquals(404, new LegendSDLCException(message, 404, cause).getStatusCode());
+        Assert.assertEquals(message, new LegendSDLCException(message).getMessage());
+        Assert.assertSame(cause, new LegendSDLCException(message, cause).getCause());
+    }
+
+    @Test
+    public void testValidateNonNull()
+    {
+        String message = "the quick brown fox jumped over the lazy dog";
+        Object someObject = new Object();
+
+        Assert.assertSame(someObject, LegendSDLCException.validateNonNull(someObject, message));
+
+        LegendSDLCException e1 = Assert.assertThrows(LegendSDLCException.class, () -> LegendSDLCException.validateNonNull(null, message));
+        Assert.assertEquals(message, e1.getMessage());
+        Assert.assertEquals(400, e1.getStatusCode());
+
+        LegendSDLCException e2 = Assert.assertThrows(message, LegendSDLCException.class, () -> LegendSDLCException.validateNonNull(null, message, 500));
+        Assert.assertEquals(message, e2.getMessage());
+        Assert.assertEquals(500, e2.getStatusCode());
+    }
+
+    @Test
+    public void testValidate_Message()
+    {
+        String message = "Cogito, ergo sum.";
+        Object validObject = new Object();
+        Object invalidObject = new Object();
+        Predicate<Object> predicate = x -> x == validObject;
+
+        Assert.assertSame(validObject, LegendSDLCException.validate(validObject, predicate, message));
+
+        LegendSDLCException e1 = Assert.assertThrows(LegendSDLCException.class, () -> LegendSDLCException.validate(invalidObject, predicate, message));
+        Assert.assertEquals(message, e1.getMessage());
+        Assert.assertEquals(400, e1.getStatusCode());
+
+        LegendSDLCException e2 = Assert.assertThrows(message, LegendSDLCException.class, () -> LegendSDLCException.validate(invalidObject, predicate, message, 501));
+        Assert.assertEquals(message, e2.getMessage());
+        Assert.assertEquals(501, e2.getStatusCode());
+    }
+
+    @Test
+    public void testValidate_MessageFunction()
+    {
+        String message = "Soon you'll be ashes or bones. A mere name at most - and even that is just a sound, an echo. The things we want in life are empty, stale, trivial.";
+        int[] messageFnCount = {0};
+        Function<Object, String> messageFn = x ->
+        {
+            messageFnCount[0]++;
+            return message;
+        };
+
+        Object validObject = new Object();
+        Object invalidObject = new Object();
+        Predicate<Object> predicate = x -> x == validObject;
+
+        Assert.assertSame(validObject, LegendSDLCException.validate(validObject, predicate, messageFn));
+        Assert.assertEquals(0, messageFnCount[0]);
+
+        LegendSDLCException e1 = Assert.assertThrows(LegendSDLCException.class, () -> LegendSDLCException.validate(invalidObject, predicate, messageFn));
+        Assert.assertEquals(message, e1.getMessage());
+        Assert.assertEquals(1, messageFnCount[0]);
+        Assert.assertEquals(400, e1.getStatusCode());
+
+        LegendSDLCException e2 = Assert.assertThrows(message, LegendSDLCException.class, () -> LegendSDLCException.validate(invalidObject, predicate, messageFn, 409));
+        Assert.assertEquals(message, e2.getMessage());
+        Assert.assertEquals(2, messageFnCount[0]);
+        Assert.assertEquals(409, e2.getStatusCode());
+    }
+}
